# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Telemetria GPS",
        "description": "API dedicada à extração e consumo de dados de rastreamento de frotas.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/esp32/dispositivos": {
            "get": {
                "tags": [
                    "ESP32 - Dispositivos"
                ],
                "summary": "Lista todos os dispositivos ESP32 cadastrados",
                "description": "Retorna a lista paginada de dispositivos com sua última telemetria.",
                "operationId": "6aa06b429585aade2ff194e6d41da475",
                "parameters": [
                    {
                        "name": "ativo",
                        "in": "query",
                        "description": "Filtrar por status (1=ativos, 0=inativos)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "por_pagina",
                        "in": "query",
                        "description": "Itens por página (padrão 15)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de dispositivos"
                    }
                }
            },
            "post": {
                "tags": [
                    "ESP32 - Dispositivos"
                ],
                "summary": "Cadastra um novo dispositivo ESP32 manualmente",
                "description": "Útil para pré-registrar um dispositivo antes de ele enviar o primeiro pacote.",
                "operationId": "34bcac34995d0ac5e2a1194d46569046",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identificador"
                                ],
                                "properties": {
                                    "identificador": {
                                        "type": "string",
                                        "example": "AA:BB:CC:DD:EE:FF"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "Sensor da Portaria"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "ESP32 instalado na portaria principal"
                                    },
                                    "ativo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dispositivo criado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "409": {
                        "description": "Identificador já cadastrado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/esp32/dispositivos/{identificador}": {
            "get": {
                "tags": [
                    "ESP32 - Dispositivos"
                ],
                "summary": "Retorna os detalhes de um dispositivo ESP32",
                "operationId": "9c3ac3e74c1da701aae146c013568d69",
                "parameters": [
                    {
                        "name": "identificador",
                        "in": "path",
                        "description": "MAC Address ou ID do dispositivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do dispositivo com última telemetria"
                    },
                    "404": {
                        "description": "Dispositivo não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "ESP32 - Dispositivos"
                ],
                "summary": "Atualiza os dados cadastrais de um dispositivo ESP32",
                "operationId": "356ced5cde9f6a5ff3eb50761d062a8e",
                "parameters": [
                    {
                        "name": "identificador",
                        "in": "path",
                        "description": "MAC Address ou ID do dispositivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Sensor Atualizado"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Localização: Galpão B"
                                    },
                                    "ativo": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dispositivo atualizado"
                    },
                    "404": {
                        "description": "Dispositivo não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ESP32 - Dispositivos"
                ],
                "summary": "Remove um dispositivo ESP32 e todo o seu histórico de telemetria",
                "operationId": "3048642ccf5b9652b60aa6bcdf2cb1a9",
                "parameters": [
                    {
                        "name": "identificador",
                        "in": "path",
                        "description": "MAC Address ou ID do dispositivo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dispositivo removido com sucesso"
                    },
                    "404": {
                        "description": "Dispositivo não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}